ADD: Aggiunto in autovoiceinteraction.js 'autoVoiceSelectMenu_creatorType'

+++ utils/autoVoice-system/events/autovoiceinteractions.js
@@ -55,8 +55,59 @@ module.exports = {
           .setDescription(language_result.selectSetup_Creator.description_embed)
           .setFooter({ text: `${language_result.selectSetup_Creator.embed_footer}`, iconURL: `${language_result.selectSetup_Creator.embed_icon_url}` })
           .setColor(0x9ba832);
-          await interaction.message.delete();
-          await interaction.reply({ embeds: [embedLog], components: [row], ephemeral: true });
+        await interaction.message.delete();
+        await interaction.reply({ embeds: [embedLog], components: [row] });
+
+      }
+
+      if (interaction.customId == 'autoVoiceSelectMenu_creatorType') {
+        let value = interaction.values;
+        await value.forEach(async v => {
+          let typeAccess;
+          switch (v) {
+            case "private":
+              typeAccess = 1;
+              break;
+            case "free":
+              typeAccess = 2;
+              break;
+
+          }
+          await runDb(`UPDATE autovoice_system_creator SET creatorType = ?`, typeAccess);
+        });
+        // RISPONDO ALL'INTERAZIONE ELIMINO IL MESSAGGIO E NE RICREO UNO NUOVO
+        const voiceNumber = await getEmoji(interaction.client, "voiceNumber");
+        const voiceUser = await getEmoji(interaction.client, "voiceUser");
+        const voiceRandom = await getEmoji(interaction.client, "voiceRandom");
+        const selectSetup_TypeAccess = new StringSelectMenuBuilder()
+          .setCustomId('autoVoiceSelectMenu_accessType')
+          .setPlaceholder(language_result.selectSetup_TypeAccess.placeholder)
+          .addOptions(
+            new StringSelectMenuOptionBuilder()
+              .setLabel(language_result.selectSetup_TypeAccess.label_numeric)
+              .setValue("numeric")
+              .setEmoji(voiceNumber.id),
+            new StringSelectMenuOptionBuilder()
+              .setLabel(language_result.selectSetup_TypeAccess.label_nickname)
+              .setValue("nickname")
+              .setEmoji(voiceUser.id),
+            new StringSelectMenuOptionBuilder()
+              .setLabel(language_result.selectSetup_TypeAccess.label_random)
+              .setValue("random")
+              .setEmoji(voiceRandom.id),
+          );
+
+        const row = new ActionRowBuilder()
+          .addComponents(selectSetup_TypeAccess);
+
+        const customEmoji = await getEmojifromUrl(interaction.client, "utilitysettings")
+        const embedLog = new EmbedBuilder()
+          .setAuthor({ name: `${language_result.selectSetup_TypeAccess.embed_title}`, iconURL: customEmoji })
+          .setDescription(language_result.selectSetup_TypeAccess.description_embed)
+          .setFooter({ text: `${language_result.selectSetup_TypeAccess.embed_footer}`, iconURL: `${language_result.selectSetup_TypeAccess.embed_icon_url}` })
+          .setColor(0x9ba832);
+        await interaction.message.delete();
+        await interaction.reply({ embeds: [embedLog], components: [row] });
 
       }
